materialsテーブルに関する処理をMaterialControllerに移動
