+++ app/Http/Controllers/RecipeController.php
@@ -38,26 +38,6 @@ class RecipeController extends Controller
         return redirect()->route('recipe.step2', ['id' => $id]);
     }
 
-    public function step2($id)
-    {
-        $recipe_id = $id;
-        return view('recipe.step2', ['recipe_id' => $recipe_id]);
-    }
-
-    public function material_store(Request $request)
-    {
-        $number = count($request->ingredients);
-        $recipe_id = $request->recipe_id;
-        for ($n = 0; $n < $number; $n++) {
-        $material = new Material;
-        $material->recipe_id = $recipe_id;
-        $material->ingredients = $request->ingredients[$n];
-        $material->quantity = $request->quantity[$n];
-        $material->save();
-        }
-        return redirect()->route('recipe.step3', ['recipe_id' => $recipe_id]);
-    }
-
     public function step3($recipe_id)
     {
         $recipe = Recipe::find($recipe_id);
@@ -127,34 +107,6 @@ class RecipeController extends Controller
         $image = str_replace('public/', '', $filePath);
         Recipe::find($recipe_id)
         ->update(['title' => $request->title, 'image' => $image, 'description'=> $request->description]);
-        $recipe = Recipe::find($recipe_id);
-        $user = Recipe::find($recipe_id)->user;
-        $materials = Recipe::find($recipe_id)->materials;
-        $processes = Recipe::find($recipe_id)->processes->sortBy('order');
-        return view('recipe.edit', compact('recipe', 'user', 'materials', 'processes'));
-    }
-
-
-    public function material_edit ($recipe_id)
-    {
-        $recipe = Recipe::find($recipe_id);
-        $user = Recipe::find($recipe_id)->user;
-        $materials = Recipe::find($recipe_id)->materials;
-        return view('recipe.material_edit', compact('recipe', 'user', 'materials'));
-    }
-
-    public function material_update (Request $request)
-    {
-        $material = Material::where('recipe_id','=',$request->recipe_id)->delete();
-        $number = count($request->ingredients);
-        $recipe_id = $request->recipe_id;
-        for ($n = 0; $n < $number; $n++) {
-        $material = new Material;
-        $material->recipe_id = $recipe_id;
-        $material->ingredients = $request->ingredients[$n];
-        $material->quantity = $request->quantity[$n];
-        $material->save();
-        }
         $recipe = Recipe::find($recipe_id);
         $user = Recipe::find($recipe_id)->user;
         $materials = Recipe::find($recipe_id)->materials;
