+++ routes/web.php
@@ -26,8 +26,8 @@ Route::get('/user', 'UserController@show');
 Route::get('/recipe', 'RecipeController@index');
 Route::get('/recipe/step1', 'RecipeController@step1');
 Route::post('/recipe/store', 'RecipeController@store')->name('recipe.store');
-Route::get('/recipe/step2/{id}', 'RecipeController@step2')->name('recipe.step2');
-Route::post('/recipe/material_store', 'RecipeController@material_store')->name('recipe.material_store');
+Route::get('/recipe/step2/{id}', 'MaterialController@step2')->name('recipe.step2');
+Route::post('/recipe/material_store', 'MaterialController@material_store')->name('recipe.material_store');
 Route::get('/recipe/step3/{recipe_id}', 'RecipeController@step3')->name('recipe.step3');
 Route::post('/recipe/process_store', 'RecipeController@process_store')->name('recipe.process_store');
 Route::get('/recipe/preview/{recipe_id}', 'RecipeController@preview')->name('recipe.preview');
@@ -36,8 +36,8 @@ Route::post('/recipe/close', 'RecipeController@close')->name('recipe.close');
 Route::get('/recipe/edit/{recipe_id}', 'RecipeController@edit');
 Route::get('/recipe/recipe_edit/{recipe_id}', 'RecipeController@recipe_edit');
 Route::patch('/recipe/recipe_update/', 'RecipeController@recipe_update')->name('recipe.update');
-Route::get('/recipe/matrial_edit/{recipe_id}', 'RecipeController@material_edit');
-Route::post('/recipe/matrial_update/', 'RecipeController@material_update')->name('recipe.material_update');
+Route::get('/recipe/matrial_edit/{recipe_id}', 'MaterialController@material_edit');
+Route::post('/recipe/matrial_update/', 'MaterialController@material_update')->name('recipe.material_update');
 Route::get('/recipe/process_edit/{recipe_id}', 'RecipeController@process_edit');
 Route::post('/recipe/process_update/', 'RecipeController@process_update')->name('recipe.process_update');
 Route::delete('/recipe/delete', 'RecipeController@delete')->name('recipe.delete');
